refactor(checkout): CHECKOUT-2951 Fix import paths
