+++ src/index.js
@@ -1,4 +1,4 @@
-export { default as createCheckoutClient } from './create-checkout-client';
-export { default as createCheckoutService } from './create-checkout-service';
+export { default as createCheckoutClient } from './checkout/create-checkout-client';
+export { default as createCheckoutService } from './checkout/create-checkout-service';
 export { default as createLanguageService } from './locale/create-language-service';
 export { createTimeout } from '@bigcommerce/request-sender';
